Adapt :font to :family to apply fonts to clients
When running Emacs as a Daemon, there are apparantly sometimes mismatches
between the loaded font and the actual font that is shown. The reason is that
the font is only loaded when loading the GUI -- which does not happen when
running an Emacs server. Changing :font to :family changes this behavior
apparantly, but please do not ask me why.