Start working on Eglot instead of LSP
As Eglot will be included in Emacs's core from version 29 onwards, and because
'lsp-mode' was giving me issues within Julia projects, I believed now
is a right time to switch to Eglot instead.

While, for now, it appears to be working in general, Eglot has the same
problems with Julia projects that use 'Pkg.add("/path/to/project")', as
the linter (flycheck or flymake) does not find references related to the
project. Still some work to do.